Optimizing Cloud Native Java, 2nd Edition - Benjamin J. Evans, James Gough Performance tuning is an experimental science, but that doesn't mean engineers should resort to guesswork and folklore to get the job done. Yet that's often the case. With this practical book, intermediate to advanced Java technologists working with complex platforms will learn how to tune Java cloud applications for performance using a quantitative, verifiable, and repeatable approach.
In response to the ubiquity of cloud computing, this updated edition of Optimizing Cloud Native Java addresses topics that are key to high performance of Java applications in the cloud. Many resources on performance tend to focus on the theory and internals of Java virtual machines, but this book discusses the low-level technical aspects within the context of performance-tuning practicalities and examines a wide range of aspects.

With this book, you will:
+  Learn how Java principles and technology make the best use of modern hardware, operating systems, and cloud stacks
+  Examine the pitfalls of measuring Java performance numbers and the drawbacks of microbenchmarking
+  Understand how to package, deploy, operate, and debug Java/JVM applications in modern cloud environments
+  Apply emerging observability approaches to obtain deep understanding of cloud native applications
+  Use Java language performance techniques including concurrent and distributed forms