FastAPI - Bill Lubanovic FastAPI is a young yet solid framework that takes advantage of newer Python features in a clean design. As its name implies, FastAPI is indeed fast, rivaling similar frameworks in languages such as Golang. With this practical book, developers familiar with Python will learn how FastAPI lets you accomplish more in less time with less code.
Author Bill Lubanovic covers the nuts and bolts of FastAPI development with how-to guides on various topics such as forms, database access, graphics, maps, and more that will take you beyond the basics. This book also includes how-to guides that will get you up to speed on RESTful APIs, data validation, authorization, and performance. With its similarities to frameworks like Flask and Django, you'll find it easy to get started with FastAPI.

Through the course of this book, you will:
+  Learn how to build web applications with FastAPI
+  Understand the differences between FastAPI, Starlette, and pydantic
+  Learn two features that set FastAPI apart: asynchronous functions and data type checking and validation
+  Examine new features of Python 3.8+, especially type annotations
+  Understand the differences between sync and async Python
+  Learn how to connect with external APIs and services