Building Applications with AI Agents - Michael Albada Generative AI has revolutionized how organizations tackle problems, accelerating the journey from concept to prototype to solution. While these applications enhance efficiency, they often require extensive planning, drafting, and revising to complete complex tasks. By combining many of these actions, AI agents offer greater autonomy and efficiency, but understanding and deploying them remains a challenge for many organizations, especially as technology and research rapidly develops.
This book is your indispensable guide through this intricate and fast-moving landscape. Author Michael Albada provides a practical and research-based approach to designing and implementing single- and multi-agent systems. It simplifies the complexities and equips you with the tools to move from concept to solution efficiently. By the end, you'll:

+  Understand the distinct features of foundation model-enabled AI agents
+  Discover the core components and design principles of AI agents
+  Explore design trade-offs and implement effective multi-agent systems
+  Design and deploy tailored AI solutions, enhancing efficiency and innovation in your field
Michael Albada is a seasoned machine learning engineer with expertise in deploying large-scale solutions for major tech firms including Uber, ServiceNow, and Microsoft. He holds degrees from Stanford University, the University of Cambridge, and Georgia Tech, specializing in machine learning.